Changed tests to run in own virtual X server
The tests are run on Xvfb, a simple X server that supports GLX,
but renders into a virtual framebuffer, of the test's requested
size.

Instead of using native image to capture the test results, the
virtual framebuffer is read directly using ImageMagick, and stored
off to a temporary file for comparison with OpenCV against
the expected result.

For some reason, the output results have a slightly different color
space to the expected images, so the acceptance criteria have changed
to 99% similarity.

Have also removed the brute force asserts on failure, which mostly
hinder debugging. Running ./execute with -v now exits cleanly, and
it's possible to catch the percentage failure with an exit code.

Executing tests individually do not use the framebuffer, instead,
they render to the user's desktop X server.

Signed-off-by: David Steele <david.steele@samsung.com>